add SettingsModal from Browse
This adds SettingsModal, a React component, to split and better
categorize related code together.
